Add support for expandlinks flag to unshorten URLs
Set the expandlinks query parameter to 1 to trigger recursive curl calls
to "unshorten" URLs from domains like bitly.com or t.co. The default is
not to do this, it must be explicitly requested.

The behaviour is to have curl issue HTTP HEAD requests on each <a> link
in the post body HTML, and follow any 30x redirects. It uses either the
last Location: header it receives or the original URL (in the event of
there being no redirection).